Recuperer le titre et les posters des films
