+++ moviecontroller.js
@@ -1,20 +1,20 @@
-//il filtre
-const moviesS = require("./movieservice")
-
-
-async function fetch_movies_title_poster(){
-
-
-    const titlesAndPosters = [];
-    for (const movie of movies) {
-        titlesAndPosters.push({
-            title: movie.title,
-            poster_path: movie.poster_path
-        });
-    }
-
-    return titlesAndPosters;
-}
-
-
-module.exports = { fetch_movies_title_poster};
+// //il filtre
+// const moviesS = require("./movieservice")
+// const fetch = require("node-fetch");
+//
+//
+// // async function fetchMovieandTitleList(){
+// //     movies=
+// //     for (const movie of movies) {
+// //         titlesAndPosters.push({
+// //             title: movie.title,
+// //             poster_path: movie.poster_path
+// //         });
+// //     }
+// //
+// //
+// // }
+//
+//
+//
+// module.exports = { fetchMovieandTitleList};
